remove BorSeparate table
